requirements.txt
langgraph
langchain
langchain_community
langchain_groq
python-dotenv
amadeus